requirements.txt

# pip install --upgrade -r requirements.txt

# googles
google-api-python-client
google-auth-httplib2
google-auth-oauthlib


### slack

slackclient
